+++ src/app/app.module.ts
@@ -8,7 +8,7 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SharedModule } from './shared/shared.module';
-import { BeginModule } from './begin/begin.module';
+import { BeginModule } from './modules/begin/begin.module';
 import { BlogModule } from './blog/blog.module';
 
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
