+++ src/app/modules/begin/begin.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 
 import { BeginRoutingModule } from './begin-routing.module';
-import { SharedModule } from '../shared/shared.module';
+import { SharedModule } from '../../shared/shared.module';
 
 import { StartComponent } from './start/start.component';
 import { WelcomeComponent } from './welcome/welcome.component';
