+++ src/app/modules/begin/start/start.component.scss
