+++ src/app/modules/begin/welcome/welcome.component.scss
