+++ src/app/modules/begin/welcome/welcome.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { PageTitleService } from '../../page-title.service';
+import { PageTitleService } from '../../../page-title.service';
 
 @Component({
   selector: 'app-welcome',
